Заменил тип double -> BigDecimal в DTO для переменных связанных с весом и объёмом доставки, так как они используются в расчете стоимости

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/BookedProductsDto.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class BookedProductsDto {
-    private double deliveryWeight;
-    private double deliveryVolume;
+    private BigDecimal deliveryWeight;
+    private BigDecimal deliveryVolume;
     private boolean fragile;
 }
